+++ include/smelt/sifting.h
@@ -9,9 +9,9 @@
 namespace forgerib {
 
 int Sift(
-	const std::vector<long long> &xia_times,
+	const std::vector<double> &xia_times,
 	const std::vector<long long> &xia_entries,
-	const std::vector<long long> &vme_times,
+	const std::vector<double> &vme_times,
 	const std::vector<long long> &vme_entries,
 	const char *output_path,
 	int group_num,
